#24 reorganised tree_dialog files
